optimisation

+++ libft/ft_min.c
@@ -6,11 +6,11 @@
 /*   By: WSeegers <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/24 12:49:03 by WSeegers          #+#    #+#             */
-/*   Updated: 2018/04/24 12:49:30 by WSeegers         ###   ########.fr       */
+/*   Updated: 2018/04/30 15:54:19 by wseegers         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long ft_min(long long nbr1, long long nbr2)
+long ft_min(long nbr1, long nbr2)
 {
 	if (nbr1 < nbr2)
 		return (nbr1);
